Split up Rails 2 and Rails 3 specific monkey patches into separate files and fix Rails 3 test support. Also rename client/action_controller_request to client/action_pack.
